Update README.md
This Script was made by Faolon, with the help of u/anonymous1184 on reddit.
I wanted to make this to be as easy to use and funcional for the community of starbase, mostly because i found the
available tool already is a pain to compile and setup for use in the game. 

all you need for this is Autohotkey and the scipt file.

if your pc is a bit slower, and has issues, try editing the ahk file and change Sleep 50 to sleep 100 for both, 
it helps slow the script down so it can be ran efficiently on some systems.



This is the script in the file:

#^v::
loop parse, Clipboard, `n, `r
{
    SendInput % "{text}" A_LoopField
     Sleep 50
    Send {Down}
     Sleep 50
    if A_Index = 20
     break
return



Autohotkey Download: https://www.autohotkey.com/download/ahk-install.exe

you can use ctrl+win+v to paste the code, ctrl+c or right clip copy is fine to copy.

V.1

Plans:Multi Clipboard (because, why not, seems useful)

Auto curser to top (Right now you have to make sure your text curser is in the number 1 box)

Overwrite previous code (because removing old code, or code to replace is a pain)

If you have any issues, questions or ideas for a feature you want, let me know!